code is refactored and devided into parts
