+++ js/full-photo-click-handler.js
@@ -1,0 +1,50 @@
+import {isEscapeKey} from './util.js';
+import {createFullPhotoCard, addCommentsInFullPhotoCard} from './full-photo-creator.js';
+
+const bigPicture = document.querySelector('.big-picture');
+const closeButton = bigPicture.querySelector('.big-picture__cancel');
+const bigPictureImage = bigPicture.querySelector('.big-picture__img img');
+const bigPictureLikes = bigPicture.querySelector('.likes-count');
+const bigPictureDescription = bigPicture.querySelector('.social__caption');
+const bigPictureCommentsValue = bigPicture.querySelector('.comments-count');
+const bigPictureCommentsList = bigPicture.querySelector('.social__comments');
+const bigPictureCommentsItem = bigPicture.querySelectorAll('.social__comment');
+const bigPictureCommentsCount = bigPicture.querySelector('.social__comment-count');
+const bigPictureCommentsLoader = bigPicture.querySelector('.comments-loader');
+const cardPictureWall = document.querySelector('.pictures');
+
+const onEscapeClick = (evt) => {
+  if (isEscapeKey(evt)) {
+    evt.preventDefault();
+    document.removeEventListener('keydown', onEscapeClick);
+    bigPicture.classList.add('hidden');
+    document.body.classList.remove('modal-open');
+  }
+};
+
+const onPictureClick = () => {
+  bigPicture.classList.remove('hidden');
+  closeButton.addEventListener('click', () => {
+    document.removeEventListener('keydown', onEscapeClick);
+    bigPicture.classList.add('hidden');
+    document.body.classList.remove('modal-open');
+  });
+  document.addEventListener('keydown', onEscapeClick);
+};
+
+const onCardClick = (evt) => {
+  if (evt.target.matches('.picture__img')) {
+    document.body.classList.add('modal-open');
+    const selectedPictureId = evt.target.dataset.pictureId;
+    createFullPhotoCard(bigPictureImage, bigPictureLikes, bigPictureDescription, bigPictureCommentsValue, bigPictureCommentsItem, selectedPictureId);
+
+    bigPictureCommentsCount.classList.add('hidden');
+    bigPictureCommentsLoader.classList.add('hidden');
+
+    bigPictureCommentsList.appendChild(addCommentsInFullPhotoCard(selectedPictureId));
+  }
+};
+
+cardPictureWall.addEventListener('click', onCardClick);
+
+export {onPictureClick};
